+++ src/test/java/com/vaadin/ui/iron/list/tests/IronListTestPage.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.iron.list.it;
+package com.vaadin.ui.iron.list.tests;
 
 import java.util.ArrayList;
 import java.util.Arrays;
